Relocate CLI-only function
